+++ include/pp/observable.hpp
@@ -1,0 +1,5 @@
+#pragma once
+
+#include <pp/observable/hot_observable.hpp>
+#include <pp/observable/observable.hpp>
+#include <pp/observable/type.hpp>
