+++ include/pp/pp.hpp
@@ -1,0 +1,6 @@
+#pragma once
+
+#include <pp/adapter.hpp>
+#include <pp/observable.hpp>
+#include <pp/observer.hpp>
+#include <pp/subject.hpp>
